+++ Software/MATLAB/test/unit/Adx/adxFixture.m
@@ -2,7 +2,7 @@ classdef adxFixture < matlab.unittest.fixtures.Fixture
     % Writes the configuration files based on environment variables, this
     % facilities unit testing in CI/CD setups as well as local  testing
     
-    % Copyright 2024 The MathWorks, Inc.
+    % Copyright 2024-2025 The MathWorks, Inc.
     
     methods
         function setup(~)
@@ -17,7 +17,7 @@ classdef adxFixture < matlab.unittest.fixtures.Fixture
                 'database',getenv('ADX_DATABASE'),...
                 'cluster',getenv('ADX_CLUSTER'),...
                 'resourceGroup',getenv('ADX_RESOURCEGROUP')...
-            ));
+            ), PrettyPrint=true);
             f = fopen(configFile,'w'); fwrite(f,json);fclose(f);
         end